% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{andes}
\alias{andes}
\title{Divergence-time estimation data for cis-trans-Andean pairs}
\format{
A data frame with three columns:
\describe{
  \item{ages}{Estimated age (in Ma) from a given rock sample}
  \item{min}{Standard deviation of the age estimate}
  \item{max}{Sample code as in Table 3.2}
}
}
\usage{
data(andes)
}
\description{
A dataset containing point estimates and uncertainty intervals of
divergence times for clade pairs east and west of the Andes,
compiled by Ballen (2020).
}
\references{
Ballen, Gustavo A. 2020. Fossil freshwater fishes and the biogeography of northern South America. 2020. PhD thesis, Museu de Zoologia, Universidade de São Paulo, São Paulo, 2020. doi:10.11606/T.38.2020.tde-06052020-181631.
}
\keyword{datasets}
