% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cynodontidae.posterior}
\alias{cynodontidae.posterior}
\title{Posterior samples for the family Cynodontidae}
\format{
A data frame named cynodontidae.posterior with six columns:
\describe{
  \item{Sample}{ID of the generation during sampling}
  \item{mrca.date.backward.Hydrolycus.}{Node ages for the node Hydrolycus}
  \item{mrca.date.backward.Cynodontidae.}{Node ages for the node Cynodontidae}
  \item{mrca.date.backward.Cynodon_Hydrolycus.}{Node ages for the node Cynodon-Hydrolycus}
  \item{mrca.date.backward.H_scomberoides.}{Node ages for the node H. scomberoides}
  \item{mrca.date.backward.Cynodon.}{Node ages for the node Cynodon}
}
}
\usage{
data(cynodontidae.posterior)
}
\description{
A data frame with posterior MCMC samples from a divergence time estimation
analysis of the family Cynodontidae using Beast2. From an original sample of 20000, these
include the last 1000 samples and only preserve the columns with node ages.
}
\references{
Ballen, G.A. and Reinales, S. 2025. tbea: tools for pre- and post-processing in Bayesian evolutionary analyses. BioRxiv https://www.biorxiv.org/content/10.1101/2024.06.18.599561.

Ballen, G.A., Moreno-Bernal, J.W. & Jaramillo, C. 2022. The fossil record of Saber-Tooth Characins (Teleostei: Characiformes: Cynodontinae), their phylogenetic relationships, and paleobiogeographical implications. Journal of Systematic Palaeontology 19:24, 1679-1692, DOI: 10.1080/14772019.2022.2070717
}
\keyword{datasets}
