% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt2newick.R
\name{tnt2newick}
\alias{tnt2newick}
\title{tnt2newick: Function for converting from TNT tree format to newick parenthetical format}
\usage{
tnt2newick(file, output, subsetting = TRUE)
}
\arguments{
\item{file}{A vector of type 'character' with the path to the original TNT tree file.}

\item{output}{A vector of type 'character' with the path to output files to contain the tree in newick format.}

\item{subsetting}{A vector of type 'logical' indicating whether subsetting (i.e., chopping at once the first and last line of the TNT tree file) should be done. Otherwise, explicit text replacements removing such lines are used.}
}
\value{
This function writes to the disk a text file containing the tree converted to newick format.
}
\description{
tnt2newick: Function for converting from TNT tree format to newick parenthetical format
}
\details{
This function has been tested for cases where only one tree is in the original tnt tree file. Please be careful with files containing multiple trees.
}
\examples{
# Convert a tree in TNT tree format to newick format
\dontrun{
tnt2newick(file = "my_TNT_tree.tre", output = "my_TNT_tree.newick")
}
}
\author{
Gustavo A. Ballen
}
