% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2nexus.R
\name{fasta2nexus}
\alias{fasta2nexus}
\title{fasta2nexus (deprecated): Function for converting molecular alignments from fasta to nexus format}
\usage{
fasta2nexus(...)
}
\arguments{
\item{...}{A placeholder for any argument that used to be in this function.}
}
\value{
This function returns an error because it has been deprecated.
}
\description{
fasta2nexus (deprecated): Function for converting molecular alignments from fasta to nexus format
}
\details{
This function was deprecated from tbea v1.0.0 onwards. For using this function please install tbea v0.5.0.
}
\author{
Gustavo A. Ballen
}
