% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbea-package.R
\docType{package}
\name{tbea-package}
\alias{tbea}
\alias{tbea-package}
\title{tbea: Tools for Pre- And Post-Processing in Bayesian Evolutionary Analyses}
\description{
Package for Bayesian inference in phylogenetics and evolution. It provides functions for prior specification in divergence time estimation using fossils as well as other kinds of data. The package provides tools for interacting with the input and output of Bayesian platforms in evolutionary biology such as BEAST2, MrBayes, RevBayes, or MCMCTree. The package implements a way to measure similarity between probability density functions in the context of comparisons between prior and posterior Bayesian densities, as well as code for calculating the combination of distributions using conflation. Tools for estimating the origination time in collections of distributions using x-intercept and stratigraphic intervals are also available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaballench/tbea}
  \item Report bugs at \url{https://github.com/gaballench/tbea/issues}
}

}
\author{
\strong{Maintainer}: Gustavo A. Ballen \email{gustavo.a.ballen@gmail.com}

Authors:
\itemize{
  \item Sandra Reinales \email{spreinalesl@gmail.com}
}

}
\keyword{internal}
