% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoFreq.R
\name{topoFreq}
\alias{topoFreq}
\title{Frequency of topologies in a tree sample}
\usage{
topoFreq(mphy, output = "index")
}
\arguments{
\item{mphy}{An object of class multiPhylo}

\item{output}{A character indicating whether the tree indices or the actual trees should be returned. Defaults to "index"}
}
\value{
A list with an element containing the the different tree clusters (as multiPhylo)
and the absolute, cumulative, and relative frequencies of each topology in the tree sample.
}
\description{
Frequency of topologies in a tree sample
}
\details{
This function can be used e.w. with a posterior sample of trees from a bayesian analysis where we
want to explore the distribution of topologies in the posterior of trees. This way we can assess topological
uncertainty in a more meaningful way than using a majority-rule consensus.
}
\examples{
# tests
set.seed(1)
trl <- ape::rmtree(10, 4)
tpf <- topoFreq(unroot(trl), output="trees")

}
